portal: replace sinhala and tamil editors with new component, update validations
